// src/graphql/__generated__/User.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: User
// ====================================================

export interface User_user {
  __typename: "User";
  displayName: string;
  id: number;
  email: string;
  avatar: string;
}

export interface User {
  user: User_user;
}